Fix builds with readonly source directories
The source directory should not be written to. So instead of writing to
a local file, then copying it to the output headers, generate it
directly to the output header directory.

This really should be using LOCAL_EXPORT_C_INCLUDE_DIRS instead of using
copy headers, but that's a much larger change.

Change-Id: I6d6a542459e587224fb7bc469a332addc8893436